Fixed finding file types of compressed files.
Changed code in jailify/metadata.py under the determine_file_type
function to include the 'fmag' object using magic.MAGIC_NONE.
This avoids the use of the .decode('UTF-8') which was throwing
an error whenever the program was given a compressed directory.

Changed code in jailify/creation.py wherever there was an
occurance of 10.2 in a regex or a string, it was changed to 10.3
to account for the newer version of FreeBSD.

Changed code in setup.py to update the version number from 1.0.4
to 1.0.5.